Datakulturarrangor.php
Klasse som representerer en Datakulturarrangor type i nominasjon
